image-scraper/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary>=2.9.10
beautifulsoup4==4.12.3
lxml>=6.0.1
